+++ store/tikv/util/failpoint.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tikv
+package util
 
 import (
 	"github.com/pingcap/failpoint"
